pyproject.toml
[tool.poetry]
name = "hobel-inperso-ieq"
version = "0.0.0"
description = "Python tools for the HOBEL InPerso IEQ project"
license = "GPL-3.0-or-later"
authors = [
	"Son Pham-Ba <son.phamba@epfl.ch>"
]
readme = "README.md"
packages = [
	{ include = "inperso" }
]

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = { version = "^3.6.2", optional = true }
pytest = { version = "^8.0.2", optional = true }
pytest-cov = { version = "^4.1.0", optional = true }
pytest-mock = { version = "^3.12.0", optional = true }
influxdb-client = "^1.41.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
requests = "^2.31.0"
pandas = "^3.0.0"

[tool.poetry.extras]
dev = [
	"pre-commit",
	"pytest",
	"pytest-cov",
	"pytest-mock",
]

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

[tool.poetry-dynamic-versioning.substitution]
files = [
	"hobel_inperso_ieq/__init__.py",
]

[tool.pytest.ini_options]
addopts = "--cov --cov-report html --cov-report term-missing --cov-fail-under 5"

[tool.poetry.scripts]
inperso-retrieve = "inperso.data_acquisition.retrieve:main"
inperso-compute-atlas-index = "inperso.atlas_index.index:main"

[tool.ruff]
line-length = 120

[tool.mypy]
python_version = "3.10"

[tool.codespell]
ignore-words-list = "Calle"
skip = "poetry.lock"
